Only run unit tests and Sonar analysis on original agent

+++ ci_scripts/ci_post_xcodebuild.sh
@@ -3,16 +3,18 @@ set -e
 
 if [ "$CI_WORKFLOW" = "Pull Request Validation" ]
 then
-  rm -rf $CI_RESULT_BUNDLE_PATH
+  if which fastlane >/dev/null; then
+    rm -rf $CI_RESULT_BUNDLE_PATH
 
-  xcodebuild \
-    -workspace "$CI_WORKSPACE/$CI_XCODE_PROJECT" \
-    -scheme "iOS Production" \
-    -destination 'platform=iOS Simulator,name=iPhone 13' \
-    -enableCodeCoverage YES \
-    -resultBundlePath $CI_RESULT_BUNDLE_PATH \
-    test-without-building
+    xcodebuild \
+      -workspace "$CI_WORKSPACE/$CI_XCODE_PROJECT" \
+      -scheme "iOS Production" \
+      -destination 'platform=iOS Simulator,name=iPhone 13' \
+      -enableCodeCoverage YES \
+      -resultBundlePath $CI_RESULT_BUNDLE_PATH \
+      test-without-building
 
-  # run in a subshell as this script changes the directory
-  (./run_sonar_analysis.sh)
+    # run in a subshell as this script changes the directory
+    (./run_sonar_analysis.sh)
+  fi
 fi
